+++ foxcross/endpoints.py
@@ -2,22 +2,15 @@ import os
 from pathlib import Path
 
 from starlette.requests import Request
-from starlette.responses import PlainTextResponse
 from starlette.templating import Jinja2Templates
 
-SCRIPT_DIR = Path(os.path.dirname(os.path.realpath(__file__)))
-templates = Jinja2Templates(directory=str(SCRIPT_DIR / "templates"))
+__location__ = Path(
+    os.path.realpath(os.path.join(os.getcwd(), os.path.dirname(__file__)))
+)
+templates = Jinja2Templates(directory=str(__location__ / "templates"))
 
 
 async def _index_endpoint(request: Request) -> Jinja2Templates.TemplateResponse:
     return templates.TemplateResponse(
         "index.html", {"request": request, "routes": request.app.routes}
     )
-
-
-async def _kubernetes_liveness_endpoint(request: Request) -> PlainTextResponse:
-    return PlainTextResponse("No problems")
-
-
-async def _kubernetes_readiness_endpoint(request: Request) -> PlainTextResponse:
-    return PlainTextResponse("No problems")