+++ foxcross/exceptions.py
@@ -1,6 +1,10 @@
-class NoServingModelsFoundError(Exception):
+class NoModelServingFoundError(Exception):
     pass
 
 
 class BadDataFormatError(Exception):
+    pass
+
+
+class TestDataPathUndefinedError(Exception):
     pass